check kill return code

+++ src/torizon/dockercompose_secondary/command_runner.cc
@@ -7,23 +7,34 @@
 // NOLINTNEXTLINE(readability-convert-member-functions-to-static)
 bool CommandRunner::run(const std::string& cmd, const api::FlowControlToken* flow_control) {
   LOG_INFO << "Running command: " << cmd;
-  boost::process::child c(cmd);
+  //boost::process::group process_group;
+  boost::process::child child_process(cmd);
 
-  while (!c.wait_for(std::chrono::milliseconds(100))) {
+  if (!child_process.valid()) {
+    LOG_WARNING << "Failed to start " << cmd;
+    return false;
+  }
+
+  while (!child_process.wait_for(std::chrono::milliseconds(100))) {
     if (flow_control != nullptr && flow_control->hasAborted()) {
       LOG_INFO << "Killing child process due to flow_control abort";
-      auto pid = c.id();
-      kill(pid, SIGTERM);
+      auto pid = child_process.id();
+      int err = kill(pid, SIGTERM);
+      if (err) {
+        LOG_WARNING << "Attempt to send SIGTERM to pid " << pid << " failed with " << strerror(errno);
+      }
       // Give it 30s to exit cleanly
-      if (!c.wait_for(std::chrono::seconds(30))) {
+      if (!child_process.wait_for(std::chrono::seconds(30))) {
         LOG_WARNING << "Process didn't respond to SIGTERM, sending SIGKILL";
-        c.terminate();
+        child_process.terminate();
       }
       return false;
     }
   }
-
-  return c.exit_code() == 0;
+  //child_process.running();
+  child_process.wait();
+  LOG_INFO << "Exit code is:" << child_process.exit_code();
+  return child_process.exit_code() == 0;
 }
 
 // NOLINTNEXTLINE(readability-convert-member-functions-to-static)
